TP4/requirements.txt
contourpy==1.2.1
cycler==0.12.1
fonttools==4.51.0
kiwisolver==1.4.5
llvmlite==0.42.0
matplotlib==3.8.4
numba==0.59.1
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pillow==10.3.0
polars==0.20.22
pyarrow==16.0.0
pyparsing==3.1.2
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
python-dateutil==2.9.0.post0
pytz==2024.1
seaborn==0.13.2
six==1.16.0
tzdata==2024.1
PyQt5==5.15.10
scikit-learn==1.5.0
python-dotenv==1.0.1